+++ src/encode/mod.rs
@@ -1,2 +1,3 @@
 pub mod gen;
+pub mod memory_pprof;
 pub mod pprof;